LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/call_indirect64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds (call_indirect64.wast:3) assert_equals: expected false but got true
PASS #3 Reinitialize the default imports
FAIL #4 Test that WebAssembly compilation succeeds (call_indirect64.wast:3) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 76: Element init_expr must produce an i32 at {loc} expected true got false
FAIL #5 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
call_indirect64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/call_indirect64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/call_indirect64.wast.js:12:3 expected true got false
FAIL #6 Test that a WebAssembly code returns a specific result (call_indirect64.wast:26) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
call_indirect64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/call_indirect64.wast.js:10:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/call_indirect64.wast.js:12:3 expected true got false
